Move webgl_generator into separate crate
