+++ webgl_generator/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "webgl_generator"
+version = "0.1.0"
+authors = [
+    "Diggory Blake",
+]
+description = "Code generators for creating bindings to the WebGL APIs."
+license = "Apache-2.0"
+documentation = "https://docs.rs/webgl_generator"
+homepage = "https://github.com/brendanzab/gl-rs/webgl_generator/"
+repository = "https://github.com/brendanzab/gl-rs/"
+readme = "README.md"
+categories = ["api-bindings", "rendering::graphics-api"]
+keywords = ["webgl"]
+
+[lib]
+name = "webgl_generator"
+path = "lib.rs"
+
+[dependencies]
+khronos_api = { version = "2.0.0", path = "../khronos_api" }
+webidl = { version = "0.4.1" }
+heck = { version = "0.3.0" }
