+++ webgl_generator/lib.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! An OpenGL bindings generator. It defines a function named `generate_bindings` which can be
+//! A WebGL bindings generator. It defines a function named `generate_bindings` which can be
 //! used to generate all constants and functions of a given OpenGL version.
 //!
 //! # Example
@@ -20,9 +20,9 @@
 //! In `build.rs`:
 //!
 //! ```no_run
-//! extern crate gl_generator;
+//! extern crate webgl_generator;
 //!
-//! use gl_generator::{Registry, Api, Profile, Fallbacks, GlobalGenerator};
+//! use webgl_generator::{Registry, Api, Profile, Fallbacks, StdwebGenerator};
 //! use std::env;
 //! use std::fs::File;
 //! use std::path::Path;
@@ -32,7 +32,7 @@
 //!     let mut file = File::create(&Path::new(&dest).join("gl_bindings.rs")).unwrap();
 //!
 //!     Registry::new(Api::Gl, (4, 5), Profile::Core, Fallbacks::All, [])
-//!         .write_bindings(GlobalGenerator, &mut file)
+//!         .write_bindings(StdwebGenerator, &mut file)
 //!         .unwrap();
 //! }
 //! ```
@@ -43,39 +43,11 @@
 //! include!(concat!(env!("OUT_DIR"), "/gl_bindings.rs"));
 //! ```
 //!
-//! # About EGL
-//!
-//! When you generate bindings for EGL, the following platform-specific types must be declared
-//!  *at the same level where you include the bindings*:
-//!
-//! - `khronos_utime_nanoseconds_t`
-//! - `khronos_uint64_t`
-//! - `khronos_ssize_t`
-//! - `EGLNativeDisplayType`
-//! - `EGLNativePixmapType`
-//! - `EGLNativeWindowType`
-//! - `EGLint`
-//! - `NativeDisplayType`
-//! - `NativePixmapType`
-//! - `NativeWindowType`
-//!
 
-#[macro_use]
-extern crate log;
-extern crate xml;
+mod webgl_generators;
+mod webgl_registry;
 
-#[cfg(feature = "unstable_generator_utils")]
-pub mod generators;
-#[cfg(not(feature = "unstable_generator_utils"))]
-mod generators;
+pub use webgl_generators::Generator;
+pub use webgl_generators::stdweb_gen::StdwebGenerator;
 
-mod registry;
-
-pub use generators::Generator;
-pub use generators::debug_struct_gen::DebugStructGenerator;
-pub use generators::global_gen::GlobalGenerator;
-pub use generators::static_gen::StaticGenerator;
-pub use generators::static_struct_gen::StaticStructGenerator;
-pub use generators::struct_gen::StructGenerator;
-
-pub use registry::*;
+pub use webgl_registry::*;
